fix(docker): improve deployment reliability and idempotent migrations
This commit addresses multiple issues discovered during Docker deployment:

1. Docker configuration improvements:
   - Hardcode DB_HOST=db and DB_USER=root in docker-compose.yml to prevent
     conflicts with local .env variable substitution
   - Add .env to .dockerignore to prevent local config from being copied
   - Modify Application.php hasEnvFile() to accept environment variables
     (for Docker deployments that don't use .env files)
   - Fix DatabaseBootstrap to always check environment variables, not just
     when .env file exists

2. Idempotent migrations:
   - Use ADD COLUMN IF NOT EXISTS instead of plain ADD COLUMN
   - Use ADD INDEX IF NOT EXISTS instead of DROP INDEX + ADD INDEX
     (MariaDB can't drop indexes used by FK constraints)
   - Remove unnecessary INT column MODIFY statements from utf8mb4 migration
     (those columns have FK constraints and don't need charset conversion)
   - Fix local_dictionaries.LdLgID column type to INT(11) to match
     languages.LgID FK target (was TINYINT causing errno 150)

Tested with fresh Docker deployment - all migrations run successfully.

+++ db/migrations/20251223_173100_add_local_dictionaries.sql
@@ -5,7 +5,7 @@
 -- Table: local_dictionaries (metadata about imported dictionaries)
 CREATE TABLE IF NOT EXISTS `local_dictionaries` (
     `LdID` INT(10) UNSIGNED NOT NULL AUTO_INCREMENT,
-    `LdLgID` TINYINT(3) UNSIGNED NOT NULL COMMENT 'Language ID this dictionary belongs to',
+    `LdLgID` INT(11) UNSIGNED NOT NULL COMMENT 'Language ID this dictionary belongs to',
     `LdName` VARCHAR(100) NOT NULL COMMENT 'Dictionary name',
     `LdDescription` VARCHAR(500) DEFAULT NULL COMMENT 'Optional description',
     `LdSourceFormat` VARCHAR(20) NOT NULL DEFAULT 'csv' COMMENT 'Original import format: csv, json, stardict',
